+++ argonaut/src/test/scala/io/fintrospect/formats/ArgonautTest.scala
@@ -68,5 +68,5 @@ class ArgonautJsonFormatTest extends JsonFormatSpec(Argonaut) {
 
   }
 
-  override val expectedJson: String = """{"string":"hello","null":null,"bigInt":12344,"object":{"field1":"aString"},"decimal":1.2,"array":["world",true],"long":2,"bool":true,"int":1}"""
+  override val expectedJson: String = """{"string":"hello","null":null,"bigInt":12344,"object":{"field1":"aString"},"decimal":1.2,"double":1.2,"array":["world",true],"long":2,"bool":true,"int":1}"""
 }