+++ spray/src/test/scala/io/fintrospect/formats/SprayJsonTest.scala
@@ -3,5 +3,5 @@ package io.fintrospect.formats
 class SprayJsonResponseBuilderTest extends JsonResponseBuilderSpec(Spray)
 
 class SprayJsonFormatTest extends JsonFormatSpec(Spray) {
-  override val expectedJson = """{"string":"hello","null":null,"bigInt":12344,"object":{"field1":"aString"},"decimal":1.2,"array":["world",true],"long":2,"bool":true,"int":1}"""
+  override val expectedJson = """{"string":"hello","null":null,"bigInt":12344,"object":{"field1":"aString"},"decimal":1.2,"double":1.2,"array":["world",true],"long":2,"bool":true,"int":1}"""
 }